pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "langgraph-supervisor"
version = "0.0.29"
description = "An implementation of a supervisor multi-agent architecture using LangGraph"
authors = [
    {name = "Vadym Barda", email = "19161700+vbarda@users.noreply.github.com "}
]
license = "MIT"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.6.0,<0.7.0",
    "langchain-core>=0.3.40",
    "langfuse>=2.0.0", # For graph observability
    "python-dotenv>=1.0.0",
    "langchain>=0.0.352",  # Supports init_chat_model
    "langchain-openai>=0.3.32"  # Required for init_chat_model
]

[dependency-groups]
test = [
    "pytest>=8.0.0",
    "ruff>=0.9.4",
    "mypy>=1.8.0",
    "pytest-socket>=0.7.0",
    "types-setuptools>=69.0.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q -v"
testpaths = [
    "tests",
]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
ignore = [
  "E501" # line-length
]


[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ty.rules]
no-matching-overload = "ignore" 
call-non-callable = "ignore"
unresolved-import = "ignore"
[tool.ty.src]
exclude = ["tests"]
